Lint improvements (#702)
Adopts the following lints across all crates:

    #![warn(
        clippy::mod_module_files,
        clippy::unwrap_used,
        missing_docs,
        rust_2018_idioms,
        unused_lifetimes,
        unused_qualifications
    )]

Also bumps the rustc version used for clippy to 1.66 and fixes all
clippy errors under the new configuration.

+++ primeorder/src/projective.rs
@@ -195,7 +195,7 @@ where
         let mut pos = C::UInt::BIT_SIZE - 4;
 
         loop {
-            let slot = (k[(pos >> 3) as usize] >> (pos & 7)) & 0xf;
+            let slot = (k[pos >> 3] >> (pos & 7)) & 0xf;
 
             let mut t = ProjectivePoint::IDENTITY;
